+++ mp_backend/settings.py
@@ -47,6 +47,7 @@ INSTALLED_APPS = [
     'rest_framework.authtoken',
     'users',
     "wallets.apps.WalletsConfig",
+    
 ]
 
 MIDDLEWARE = [