+++ wallets/serializers.py
@@ -1,5 +1,6 @@
 from rest_framework import serializers
 from wallets.models import Wallet
+from decimal import Decimal
 
 class WalletSerializer(serializers.ModelSerializer):
     class Meta:
@@ -9,3 +10,10 @@ class WalletSerializer(serializers.ModelSerializer):
             "balance",
             "is_active",
         )
+
+class DepositSerializer(serializers.Serializer):
+    amount = serializers.DecimalField(
+        max_digits=12,
+        decimal_places=2,
+        min_value=Decimal("0.01")
+    )